# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.AccountSummary:
    properties:
      balance:
        description: Can be string or decimal.Decimal depending on model version
      currency:
        $ref: '#/definitions/dtos.CurrencySummary'
      id:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
...
  dtos.CategorySummary:
    properties:
      id:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
...
  dtos.CreateTransactionRequest:
    properties:
      account_from_id:
        type: integer
      account_to_id:
        type: integer
      amount:
        type: number
      category_id:
        type: integer
      description:
        maxLength: 255
        minLength: 1
        type: string
      exchange_rate:
        type: number
      notes:
        maxLength: 1000
        type: string
      transaction_date:
        description: ISO 8601 format
        type: string
      type:
        enum:
        - INCOME
        - EXPENSE
        - TRANSFER
        type: string
    required:
    - account_from_id
    - amount
    - description
    - transaction_date
    - type
    type: object
...
  dtos.CurrencySummary:
    properties:
      code:
        type: string
      id:
        type: integer
      symbol:
        type: string
    type: object
...
  dtos.ErrorResponse:
    properties:
      details:
        example: additional error details
        type: string
      error:
        example: error message
        type: string
    type: object
...
  dtos.SuccessResponse:
    properties:
      message:
        example: operation completed successfully
        type: string
    type: object
  dtos.TransactionListResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/dtos.TransactionResponse'
        type: array
    type: object
  dtos.TransactionResponse:
    properties:
      account_from:
        allOf:
        - $ref: '#/definitions/dtos.AccountSummary'
        description: Relationships
      account_to:
        $ref: '#/definitions/dtos.AccountSummary'
      amount:
        type: number
      amount_in_usd:
        type: number
      category:
        $ref: '#/definitions/dtos.CategorySummary'
      created_at:
        type: string
      description:
        type: string
      exchange_rate:
        type: number
      id:
        type: integer
      is_reconciled:
        type: boolean
      notes:
        type: string
      transaction_date:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  dtos.UpdateTransactionRequest:
    properties:
      amount:
        type: number
      description:
        maxLength: 255
        minLength: 1
        type: string
      is_reconciled:
        type: boolean
      notes:
        maxLength: 1000
        type: string
      transaction_date:
        description: ISO 8601 format
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@arabella.app
    name: Arabella Support
  description: API REST para el sistema de gestión financiera personal Arabella. Implementa
    contabilidad de doble partida, cálculo de Runway y soporte multi-moneda.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Arabella Financial OS API
  version: 1.0.0
paths:
...
  /transactions:
    get:
      description: Obtiene una lista paginada de transacciones del usuario autenticado
        con filtros opcionales. Cada transacción pasa por el Motor Contable de doble
        partida
      parameters:
      - description: Tipo de transacción (INCOME, EXPENSE, TRANSFER)
        in: query
        name: type
        type: string
      - description: Filtrar por ID de cuenta
        in: query
        name: account_id
        type: integer
      - description: Filtrar por ID de categoría
        in: query
        name: category_id
        type: integer
      - description: 'Número de página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Elementos por página (default: 20, máx: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de transacciones paginada
          schema:
            $ref: '#/definitions/dtos.TransactionListResponse'
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar transacciones
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      description: 'Crea una nueva transacción procesándola a través del Motor Contable
        de doble partida. Genera automáticamente los asientos de débito y crédito
        y actualiza los saldos de las cuentas involucradas. Tipos: INCOME (requiere
        category_id), EXPENSE (requiere category_id), TRANSFER (requiere account_to_id)'
      parameters:
      - description: Datos de la transacción
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Transacción creada y contabilizada exitosamente
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.TransactionResponse'
              message:
                type: string
            type: object
        "400":
          description: Datos inválidos o regla de negocio violada
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear transacción
      tags:
      - Transactions
  /transactions/{id}:
    delete:
      description: Elimina una transacción creando asientos de reversión en el diario
        contable (DEBIT↔CREDIT invertidos). Nunca borra registros del diario; garantiza
        la integridad del historial contable. Los saldos de las cuentas se restauran
        automáticamente
      parameters:
      - description: ID de la transacción a revertir
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transacción revertida exitosamente
          schema:
            $ref: '#/definitions/dtos.SuccessResponse'
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar transacción (reversión contable)
      tags:
      - Transactions
    get:
      description: Obtiene el detalle completo de una transacción incluyendo sus asientos
        contables relacionados (account_from, account_to, categoría)
      parameters:
      - description: ID de la transacción
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detalle de la transacción
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.TransactionResponse'
            type: object
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Transacción no encontrada
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener transacción por ID
      tags:
      - Transactions
    put:
      consumes:
      - application/json
      description: Actualiza los campos editables de una transacción existente (descripción,
        notas, fecha, estado de conciliación). El monto y las cuentas no pueden modificarse
        directamente; para ello, elimina la transacción y crea una nueva
      parameters:
      - description: ID de la transacción
        in: path
        name: id
        required: true
        type: integer
      - description: Campos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Transacción actualizada exitosamente
          schema:
            $ref: '#/definitions/dtos.SuccessResponse'
        "400":
          description: ID o datos inválidos
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar transacción
      tags:
      - Transactions
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Escribe "Bearer" seguido de un espacio y tu JWT de acceso. Ejemplo:
      "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"